blog/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-el-pagination==4.1.2
Markdown==3.7
pillow==11.0.0
setuptools==75.6.0
sqlparse==0.5.3
